Move `ThemeProvider`
